requirements.txt
flask==3.1.0
gunicorn==23.0.0
elevenlabs==1.50.5
google-adk==0.5.0
google-genai==1.9.0
google-cloud-secret-manager==2.22.0
google-cloud-logging==3.11.4
requests==2.32.3
